requirements.txt
tensorflow
numpy
streamlit
pillow
scikit-learn
tqdm
